+++ packages/ceti-tag-data-capture/src/sensors/pressure_temperature.c
@@ -67,7 +67,7 @@ void *pressureTemperature_thread(void *paramPtr) {
 
   // Main loop while application is running.
   CETI_LOG("Starting loop to periodically acquire data");
-  int64_t global_time_us;
+  long long global_time_us;
   int rtc_count;
   int64_t polling_sleep_duration_us;
   g_pressureTemperature_thread_is_running = 1;
